Move RegularGraph into regular.cc and delete constraint_state.hh
